+++ src/bvals/comms/bnd_info.hpp
@@ -206,7 +206,7 @@ struct BvarsCache_t {
   // auto &operator[](BoundaryType boundType) { return
   // caches[static_cast<int>(boundType)]; }
   void clear() {
-    for (int i = 0; i < caches.size(); ++i)
+    for (size_t i = 0; i < caches.size(); ++i)
       caches[i].clear();
   }
 };
