+++ src/pack/sparse_pack.hpp
@@ -82,7 +82,7 @@ class SparsePack : public SparsePackBase {
 
   // Returns the total number of vars/components in pack
   KOKKOS_FORCEINLINE_FUNCTION
-  int GetSize() const { return size_; }
+  size_t GetSize() const { return size_; }
 
   KOKKOS_INLINE_FUNCTION
   const Coordinates_t &GetCoordinates(const int b = 0) const { return coords_(b)(); }
